+++ Dev/ObjectiveCAppExample/AppDelegate.m
@@ -17,24 +17,22 @@
 
 - (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions {
 
+    InstanaSetupOptions* options = nil;
+
+    //note: explicitly get user permission before setting enableCrashReporting to true
+//    options = [[InstanaSetupOptions alloc] initWithHttpCaptureConfig: 0
+//                                           collectionEnabled: true
+//                                           enableCrashReporting: true
+//                                           suspendReportingOnLowBattery: true
+//                                           suspendReportingOnCellular: false
+//                                           slowSendInterval: 0.0
+//                                           usiRefreshTimeIntervalInHrs: -1
+//                                           autoCaptureScreenNames: true
+//                                           debugAllScreenNames: false];
+
     (void)[Instana setupWithKey: @"INSTANA_REPORTING_KEY"
              reportingURL: [NSURL URLWithString: @"INSTANA_REPORTING_URL"]
-                  options: nil];
-
-//    NSArray<NSString *> *viewControllerClasses = @[@"ViewController"];
-//    //note: explicitly get user permission before setting enableCrashReporting to true
-//    InstanaSetupOptions* options = [[InstanaSetupOptions alloc] initWithHttpCaptureConfig: 0
-//                                                                  collectionEnabled: true
-//                                                               enableCrashReporting: true
-//                                                             suspendReportingOnLowBattery: true
-//                                                               suspendReportingOnCellular: false
-//                                                                   slowSendInterval: 0.0
-//                                                              usiRefreshTimeIntervalInHrs: -1
-//                                                                autoCaptureScreenNames: 2
-//                                                            autoViewCaptureAllowedClasses: viewControllerClasses];
-//    (void)[Instana setupWithKey: @"INSTANA_REPORTING_KEY"
-//             reportingURL: [NSURL URLWithString: @"INSTANA_REPORTING_URL"]
-//                  options: options];
+                  options: options];
 
     NSURL* url = [NSURL URLWithString: @"https://www.ibm.com/jp-ja"];
     NSURLRequest* request = [NSURLRequest requestWithURL: url];